backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
numpy==1.24.3
scipy==1.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-dotenv==1.0.0 